fix: more tweaks to the client for strange invalid token issue
