+++ app/src/bcsc-theme/api/client.ts
@@ -164,6 +164,10 @@ class BCSCApiClient {
         throw new Error('Refresh token expired')
       }
 
+      if (!this.isTokenExpired(this.tokens.access_token)) {
+        return this.tokens
+      }
+
       // access token is expired, fetch new tokens using refresh token
       return this.getTokensForRefreshToken(this.tokens.refresh_token)
     })
@@ -188,13 +192,9 @@ class BCSCApiClient {
       return config
     }
 
-    if (!this.tokens || !this.isTokenExpired(this.tokens.access_token)) {
-      this.tokens = await this.fetchAccessToken()
-    }
+    this.tokens = await this.fetchAccessToken()
 
-    if (this.tokens) {
-      config.headers.set('Authorization', `Bearer ${this.tokens.access_token}`)
-    }
+    config.headers.set('Authorization', `Bearer ${this.tokens.access_token}`)
 
     return config
   }
