+++ app/src/bcsc-theme/navigators/MainStack.tsx
@@ -1,7 +1,7 @@
 import { HelpCentreUrl } from '@/constants'
-import { testIdWithKey, useDefaultStackOptions, useStore, useTheme, useTour } from '@bifold/core'
+import { testIdWithKey, useDefaultStackOptions, useTheme, useTour } from '@bifold/core'
 import { createStackNavigator } from '@react-navigation/stack'
-import { useEffect, useMemo, useRef } from 'react'
+import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 import { View } from 'react-native'
 import createHelpHeaderButton from '../components/HelpHeaderButton'
@@ -18,45 +18,16 @@ import WebViewScreen from '../features/webview/WebViewScreen'
 import { BCSCRootStackParams, BCSCScreens, BCSCStacks } from '../types/navigators'
 import BCSCTabStack from './TabStack'
 import EditNicknameScreen from '../features/account/EditNicknameScreen'
-import { DeviceCountStartupCheck, runStartupChecks, ServerStatusStartupCheck } from '@/services/StartupChecks'
-import useApi from '../api/hooks/useApi'
+import { useSystemChecks } from '../hooks/useStartupSystemChecks'
 
 const MainStack: React.FC = () => {
-  const api = useApi()
   const { currentStep } = useTour()
   const theme = useTheme()
   const { t } = useTranslation()
   const Stack = createStackNavigator<BCSCRootStackParams>()
   const hideElements = useMemo(() => (currentStep === undefined ? 'auto' : 'no-hide-descendants'), [currentStep])
   const defaultStackOptions = useDefaultStackOptions(theme)
-  const [, dispatch] = useStore()
-  const startupCheckRef = useRef(false)
-
-  // TODO (MD): Move into its own file (useMainStackStartupChecks or useStartupChecks)
-  useEffect(() => {
-    const asyncEffect = async () => {
-      if (startupCheckRef.current) {
-        return
-      }
-
-      await runStartupChecks([
-        new ServerStatusStartupCheck({
-          dispatch,
-          bannerTitle: t('StartupChecks.ServerStatusBannerTitle'),
-          getServerStatus: () => api.config.getServerStatus(),
-        }),
-        new DeviceCountStartupCheck({
-          dispatch,
-          bannerTitle: t('StartupChecks.DeviceLimitBannerTitle'),
-          getIdToken: () => api.token.getCachedIdTokenMetadata({ refreshCache: true }),
-        }),
-      ])
-
-      startupCheckRef.current = true
-    }
-
-    asyncEffect()
-  }, [api.config, api.token, dispatch, t])
+  useSystemChecks()
 
   return (
     <View style={{ flex: 1 }} importantForAccessibility={hideElements}>
